metadata import file handling fix and unload of aggregates

+++ src/main/java/com/revealprecision/revealserver/api/v1/dto/factory/LocationFormDataAggregateEventFactory.java
@@ -5,7 +5,7 @@ import com.revealprecision.revealserver.messaging.message.LocationFormDataAggreg
 public class LocationFormDataAggregateEventFactory {
 
   public static LocationFormDataAggregateEvent getEntity(LocationFormDataAggregateEvent v,
-      Long increment, Long averageCounter, String tagValue) {
+      Double increment, Long averageCounter, String tagValue) {
 
     return LocationFormDataAggregateEvent.builder()
         .plan(v.getPlan())
